Better __main__.py solution.
